Payment domain and public contracts for Nets payment provider
