+++ Api/BccPay.Core.Contracts/Requests/Nodes/NotificationRequest.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace BccPay.Core.Contracts.Requests.Nodes
+{
+    public class NotificationRequest
+    {
+        public List<WebHoorRequest> WebHoorRequests { get; set; }
+    }
+}
