+++ Api/BccPay.Core.Contracts/Requests/Nodes/OrderRequest.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace BccPay.Core.Contracts.Requests.Nodes
+{
+    public class OrderRequest
+    {
+        public List<ItemRequest> Items { get; set; }
+        public int Amount { get; set; }
+        public string Currency { get; set; }
+    }
+}
