+++ Api/BccPay.Core.Domain/BccPay.Core.Domain.csproj
@@ -13,4 +13,8 @@
     <PackageReference Include="RavenDB.Client" Version="5.2.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\BccPay.Core.Enums\BccPay.Core.Enums.csproj" />
+  </ItemGroup>
+
 </Project>
