+++ Api/BccPay.Core.Infrastructure/PaymentModels/NetsNodes/Address.cs
@@ -1,0 +1,12 @@
+﻿namespace BccPay.Core.Infrastructure.PaymentModels.NetsNodes
+{
+    public class Address
+    {
+        public string AddressLine1 { get; set; }
+        public string AddressLine2 { get; set; }
+        public string ReceiverLine { get; set; }
+        public string PostalCode { get; set; }
+        public string City { get; set; }
+        public string Country { get; set; }
+    }
+}
