+++ Api/BccPay.Core.Infrastructure/PaymentModels/NetsNodes/CheckoutOnRetrieve.cs
@@ -1,0 +1,8 @@
+﻿namespace BccPay.Core.Infrastructure.PaymentModels.NetsNodes
+{
+    public class CheckoutOnRetrieve
+    {
+        public string Url { get; set; }
+        public string CancelUrl { get; set; }
+    }
+}
