+++ Api/BccPay.Core.Infrastructure/PaymentModels/NetsNodes/Consumer.cs
@@ -1,0 +1,10 @@
+﻿namespace BccPay.Core.Infrastructure.PaymentModels.NetsNodes
+{
+    public class Consumer
+    {
+        public Address ShippingAddress { get; set; }
+        public Company Company { get; set; }
+        public PrivatePerson PrivatePerson { get; set; }
+        public Address BillingAddress { get; set; }
+    }
+}
