+++ Api/BccPay.Core.Infrastructure/PaymentModels/NetsNodes/ContactDetails.cs
@@ -1,0 +1,10 @@
+﻿namespace BccPay.Core.Infrastructure.PaymentModels.NetsNodes
+{
+    public class ContactDetails
+    {
+        public string FirstName { get; set; }
+        public string LastName { get; set; }
+        public string Email { get; set; }
+        public PhoneNumber PhoneNumber { get; set; }
+    }
+}
