+++ Api/BccPay.Core.Infrastructure/PaymentModels/NetsNodes/Fee.cs
@@ -1,0 +1,15 @@
+﻿namespace BccPay.Core.Infrastructure.PaymentModels.NetsNodes
+{
+    public class Fee
+    {
+        public string Reference { get; set; }
+        public string Name { get; set; }
+        public int Quantity { get; set; }
+        public string Unit { get; set; }
+        public decimal UnitPrice { get; set; }
+        public int TaxRate { get; set; }
+        public decimal TaxAmount { get; set; }
+        public decimal GrossTotalAmount { get; set; }
+        public decimal NetTotalAmount { get; set; }
+    }
+}
