+++ Api/BccPay.Core.Infrastructure/PaymentModels/NetsNodes/OrderDetails.cs
@@ -1,0 +1,9 @@
+﻿namespace BccPay.Core.Infrastructure.PaymentModels.NetsNodes
+{
+    public class OrderDetails
+    {
+        public int Amount { get; set; }
+        public string Currency { get; set; }
+        public string Reference { get; set; }
+    }
+}
