+++ Api/BccPay.Core.Infrastructure/PaymentModels/NetsNodes/PaymentMethod.cs
@@ -1,0 +1,8 @@
+﻿namespace BccPay.Core.Infrastructure.PaymentModels.NetsNodes
+{
+    public class PaymentMethod
+    {
+        public string Name { get; set; }
+        public Fee Fee { get; set; }
+    }
+}
