+++ Api/BccPay.Core.Infrastructure/PaymentModels/NetsNodes/PhoneNumber.cs
@@ -1,0 +1,8 @@
+﻿namespace BccPay.Core.Infrastructure.PaymentModels.NetsNodes
+{
+    public class PhoneNumber
+    {
+        public string Prefix { get; set; }
+        public string Number { get; set; }
+    }
+}
