+++ Api/BccPay.Core.Infrastructure/PaymentModels/NetsNodes/Subscription.cs
@@ -1,0 +1,7 @@
+﻿namespace BccPay.Core.Infrastructure.PaymentModels.NetsNodes
+{
+    public class Subscription
+    {
+        public string Id { get; set; }
+    }
+}
