+++ Api/BccPay.Core.Sample/BccPay.Core.Sample.csproj
@@ -13,6 +13,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\BccPay.Core.Contracts\BccPay.Core.Contracts.csproj" />
     <ProjectReference Include="..\BccPay.Core.Domain\BccPay.Core.Domain.csproj" />
     <ProjectReference Include="..\BccPay.Core.Infrastructure\BccPay.Core.Infrastructure.csproj" />
   </ItemGroup>
